# frontend/docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "GalaxyRend Backend API",
        "description": "Backend service for GalaxyRend decentralized computing platform",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/auth/challenge": {
            "post": {
                "tags": [
                    "authentication"
                ],
                "summary": "Get Auth Challenge",
                "description": "Generate authentication challenge for wallet signing.",
                "operationId": "get_auth_challenge_api_v1_auth_challenge_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChallengeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChallengeResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/authenticate": {
            "post": {
                "tags": [
                    "authentication"
                ],
                "summary": "Authenticate User",
                "description": "Authenticate user with signed challenge. Creates user if doesn't exist.",
                "operationId": "authenticate_user_api_v1_auth_authenticate_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/worker-auth": {
            "post": {
                "tags": [
                    "authentication"
                ],
                "summary": "Authenticate Worker",
                "description": "Authenticate worker with signed challenge. Creates worker if doesn't exist.",
                "operationId": "authenticate_worker_api_v1_auth_worker_auth_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "tags": [
                    "authentication"
                ],
                "summary": "Refresh Token",
                "description": "Refresh access token using refresh token.",
                "operationId": "refresh_token_api_v1_auth_refresh_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RefreshRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RefreshResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/verify-token": {
            "post": {
                "tags": [
                    "authentication"
                ],
                "summary": "Verify Token",
                "description": "Verify token validity.",
                "operationId": "verify_token_api_v1_auth_verify_token_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifyTokenRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VerifyTokenResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "authentication"
                ],
                "summary": "Logout",
                "description": "Logout worker (invalidate tokens).",
                "operationId": "logout_api_v1_auth_logout_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LogoutRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "tags": [
                    "authentication"
                ],
                "summary": "Get Current Worker",
                "description": "Get current authenticated worker information.",
                "operationId": "get_current_worker_api_v1_auth_me_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkerResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/auth/me/user": {
            "get": {
                "tags": [
                    "authentication"
                ],
                "summary": "Get Current User Info",
                "description": "Get current authenticated user information.",
                "operationId": "get_current_user_info_api_v1_auth_me_user_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuthRequest": {
                "properties": {
                    "address": {
                        "type": "string",
                        "title": "Address",
                        "description": "StarkNet address"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "Original challenge message"
                    },
                    "signature": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Signature",
                        "description": "Signature components"
                    },
                    "timestamp": {
                        "type": "integer",
                        "title": "Timestamp",
                        "description": "Challenge timestamp"
                    }
                },
                "type": "object",
                "required": [
                    "address",
                    "message",
                    "signature",
                    "timestamp"
                ],
                "title": "AuthRequest",
                "description": "Request for authentication with signed challenge."
            },
            "AuthResponse": {
                "properties": {
                    "access_token": {
                        "type": "string",
                        "title": "Access Token",
                        "description": "JWT access token"
                    },
                    "refresh_token": {
                        "type": "string",
                        "title": "Refresh Token",
                        "description": "JWT refresh token"
                    },
                    "token_type": {
                        "type": "string",
                        "title": "Token Type",
                        "description": "Token type",
                        "default": "bearer"
                    },
                    "expires_in": {
                        "type": "integer",
                        "title": "Expires In",
                        "description": "Token expiration time in seconds"
                    },
                    "user": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/UserResponse"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "description": "Authenticated user information"
                    },
                    "worker": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/WorkerResponse"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "description": "Authenticated worker information"
                    }
                },
                "type": "object",
                "required": [
                    "access_token",
                    "refresh_token",
                    "expires_in"
                ],
                "title": "AuthResponse",
                "description": "Response after successful authentication."
            },
            "ChallengeRequest": {
                "properties": {
                    "address": {
                        "type": "string",
                        "title": "Address",
                        "description": "StarkNet address"
                    }
                },
                "type": "object",
                "required": [
                    "address"
                ],
                "title": "ChallengeRequest",
                "description": "Request for authentication challenge."
            },
            "ChallengeResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "Message to be signed"
                    },
                    "timestamp": {
                        "type": "string",
                        "title": "Timestamp",
                        "description": "Challenge timestamp"
                    },
                    "nonce": {
                        "type": "string",
                        "title": "Nonce",
                        "description": "Random nonce"
                    },
                    "expires_at": {
                        "type": "string",
                        "title": "Expires At",
                        "description": "Challenge expiration timestamp"
                    }
                },
                "type": "object",
                "required": [
                    "message",
                    "timestamp",
                    "nonce",
                    "expires_at"
                ],
                "title": "ChallengeResponse",
                "description": "Response containing authentication challenge."
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "LogoutRequest": {
                "properties": {
                    "refresh_token": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Refresh Token",
                        "description": "Refresh token to invalidate"
                    }
                },
                "type": "object",
                "title": "LogoutRequest",
                "description": "Request to logout (invalidate tokens)."
            },
            "RefreshRequest": {
                "properties": {
                    "refresh_token": {
                        "type": "string",
                        "title": "Refresh Token",
                        "description": "Valid refresh token"
                    }
                },
                "type": "object",
                "required": [
                    "refresh_token"
                ],
                "title": "RefreshRequest",
                "description": "Request to refresh access token."
            },
            "RefreshResponse": {
                "properties": {
                    "access_token": {
                        "type": "string",
                        "title": "Access Token",
                        "description": "New JWT access token"
                    },
                    "token_type": {
                        "type": "string",
                        "title": "Token Type",
                        "description": "Token type",
                        "default": "bearer"
                    },
                    "expires_in": {
                        "type": "integer",
                        "title": "Expires In",
                        "description": "Token expiration time in seconds"
                    }
                },
                "type": "object",
                "required": [
                    "access_token",
                    "expires_in"
                ],
                "title": "RefreshResponse",
                "description": "Response with new access token."
            },
...
            "TokenPayload": {
                "properties": {
                    "sub": {
                        "type": "string",
                        "title": "Sub",
                        "description": "Subject (worker address)"
                    },
                    "role": {
                        "type": "string",
                        "title": "Role",
                        "description": "User role",
                        "default": "worker"
                    },
                    "exp": {
                        "type": "integer",
                        "title": "Exp",
                        "description": "Expiration timestamp"
                    },
                    "iat": {
                        "type": "integer",
                        "title": "Iat",
                        "description": "Issued at timestamp"
                    },
                    "type": {
                        "type": "string",
                        "title": "Type",
                        "description": "Token type (access/refresh)"
                    }
                },
                "type": "object",
                "required": [
                    "sub",
                    "exp",
                    "iat",
                    "type"
                ],
                "title": "TokenPayload",
                "description": "JWT token payload structure."
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
            "VerifyTokenRequest": {
                "properties": {
                    "token": {
                        "type": "string",
                        "title": "Token",
                        "description": "Token to verify"
                    }
                },
                "type": "object",
                "required": [
                    "token"
                ],
                "title": "VerifyTokenRequest",
                "description": "Request to verify token validity."
            },
            "VerifyTokenResponse": {
                "properties": {
                    "valid": {
                        "type": "boolean",
                        "title": "Valid",
                        "description": "Whether token is valid"
                    },
                    "payload": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/TokenPayload"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "description": "Token payload if valid"
                    },
                    "error": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error",
                        "description": "Error message if invalid"
                    }
                },
                "type": "object",
                "required": [
                    "valid"
                ],
                "title": "VerifyTokenResponse",
                "description": "Response with token verification result."
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}